Create separate pnm module
