+++ Cargo.toml
@@ -51,13 +51,13 @@ num-complex = "0.1.32"
 glob = "0.2.10"
 
 [features]
-default = ["gif_codec", "jpeg", "ico", "png_codec", "ppm", "tga", "tiff", "webp", "bmp", "hdr"]
+default = ["gif_codec", "jpeg", "ico", "png_codec", "pnm", "tga", "tiff", "webp", "bmp", "hdr"]
 
 gif_codec = ["gif"]
 ico = ["bmp", "png_codec"]
 jpeg = ["jpeg-decoder"]
 png_codec = ["png"]
-ppm = []
+pnm = []
 tga = []
 tiff = []
 webp = []
